+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* Setting `ShareDefaultModule = true` in a `@meta` block makes all unnamed `@example`, `@repl` and `@setup` blocks share the same sandbox module, which makes writing tutorial-style documents more practical. ([#2656])
 * `makedocs` and `doctest` now accept regex/substitution pairs in `doctestfilters`. ([#2360], [#2619])
 * Doctests now can have `teardown` code in addition to `setup` code. ([#2058], [#2566], [#2577])
 * The search UI will again put the query in the site URL, allowing deep links to searches to be shared. ([#2543], [#2646])